accountdetail seperate to UserDetailController
