+++ backend/src/main/java/com/app/ecommerce/services/IAccountDetailServices.java
@@ -11,6 +11,8 @@ public interface IAccountDetailServices {
 
     public List<AccountDetail> getAccountDetailActive();
 
+    public List<AccountDetail> getAccountDetailNotActive();
+
     public AccountDetail saveAccountDetail(CreateAccountDetailDTO accountDetail);
 
     public AccountDetail updateAccountDetail(String id, UpdateAccountDetailDTO accountDetail);
